sm6150-common: Decommonize vendor security patch level
This common tree is updated further than the EOL state of
some devices now. For EOL devices we can not update the
device specific proprietary vendor blobs. Since the vendor
security patch level should be set according to the least
updated components in vendor this differs betweeen devices,
hence decommonize it so that devices with up-to-date stock
can update it.

Change-Id: I320b2f93f7b78d01393259662450f328bbef5484